refactor: move resolve options from config to persistentCache
